+++ rscanner/src/execute/icmp/receive.rs
@@ -8,6 +8,7 @@ use pnet::packet::Packet;
 use tokio::sync::OnceCell;
 
 use super::common;
+use super::receive;
 
 static RECEIVE_PACKETS: OnceCell<Arc<Mutex<HashSet<IpAddr>>>> = OnceCell::const_new();
 
@@ -31,8 +32,8 @@ pub async fn receive_packets() -> anyhow::Result<()> {
     loop {
         tokio::time::sleep(Duration::from_nanos(1)).await;
         if let Ok(Some((packet, addr))) = iter.next_with_timeout(Duration::from_secs(1)) {
-            if is_addr_received(&addr).await {
-                continue;
+            if receive::is_addr_received(&addr).await {
+                return Ok(());
             }
             if let Some(reply_packet) = EchoReplyPacket::new(packet.packet()) {
                 if reply_packet.get_icmp_type() == IcmpTypes::EchoReply {